Add database struct with parse fn
